+++ test/unit/lib/inputs.js
@@ -1,8 +1,86 @@
 import _ from "lodash";
 
-import sharedInputs from "../../../../lib/shared/inputs.js";
+import inputs from "../../../lib/inputs.js";
 
-describe("Shared inputs", () => {
+describe("Inputs", () => {
+  describe("packageName", () => {
+    beforeEach(function () {
+      this.generator = {
+        destinationPath: sinon.fake(),
+        fs: { read: sinon.stub() },
+      };
+      this.input = inputs.packageName.create(this.generator);
+      this.cwd = sinon.stub(process, "cwd");
+    });
+
+    afterEach(function () {
+      this.cwd.restore();
+    });
+
+    it('defaults to read "name" from pyproject.toml', async function () {
+      this.generator.fs.read.returns(`
+        [tool.poetry]
+        name = "ganryu"
+      `);
+
+      const promptDefault = await this.input.asPrompt().default();
+
+      promptDefault.should.equal("ganryu");
+      this.generator.fs.read.should.have.been.calledOnce;
+      process.cwd.should.not.have.been.called;
+    });
+
+    it("defaults to the current working directory name when reading pyproject.toml is undefined", async function () {
+      this.generator.fs.read.returns("");
+      process.cwd.returns("/an/absolute/path/to/here");
+
+      const promptDefault = await this.input.asPrompt().default();
+
+      promptDefault.should.equal("here");
+      process.cwd.should.have.been.calledOnce;
+    });
+
+    it("defaults to null if the current working directory is not a valid python package name", async function () {
+      this.generator.fs.read.returns("");
+      process.cwd.returns("/an/absolute/path/to-here");
+
+      const promptDefault = await this.input.asPrompt().default();
+
+      should.equal(promptDefault, null);
+      process.cwd.should.have.been.calledOnce;
+    });
+  });
+
+  describe("packageVersion", () => {
+    beforeEach(function () {
+      this.generator = {
+        destinationPath: sinon.fake(),
+        fs: { read: sinon.stub() },
+      };
+      this.input = inputs.packageVersion.create(this.generator);
+    });
+
+    it('defaults to read "version" from pyproject.toml', async function () {
+      this.generator.fs.read.returns(`
+        [tool.poetry]
+        version = "0.0.7"
+      `);
+
+      const promptDefault = await this.input.asPrompt().default();
+
+      promptDefault.should.equal("0.0.7");
+      this.generator.fs.read.should.have.been.calledOnce;
+    });
+
+    it('defaults to the "0.0.0" when reading pyproject.toml is undefined', async function () {
+      this.generator.fs.read.returns("");
+
+      const promptDefault = await this.input.asPrompt().default();
+
+      promptDefault.should.equal("0.0.0");
+    });
+  });
+
   describe("repository", () => {
     beforeEach(function () {
       this.generator = {
@@ -10,7 +88,7 @@ describe("Shared inputs", () => {
         fs: { read: sinon.stub() },
         _queryGitOriginUrl: sinon.stub(),
       };
-      this.input = sharedInputs.repository.create(this.generator);
+      this.input = inputs.repository.create(this.generator);
     });
 
     it('defaults to read "repository" from pyproject.toml', async function () {
@@ -66,7 +144,7 @@ describe("Shared inputs", () => {
           fs: { read: sinon.stub() },
           user: { git: { [gitConfigName]: this.gitConfigStub } },
         };
-        this.input = sharedInputs[inputName].create(this.generator);
+        this.input = inputs[inputName].create(this.generator);
       });
 
       it('defaults to read "authors" in pyproject.toml', async function () {